agentic_obfuscator_deobfuscator/pyproject.toml
[project]
name = "obfuscation_deobfuscation_crew"
version = "0.1.0"
description = "Obfuscation-deobfuscation crew using CrewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "bandit>=1.8.3",
    "crewai[tools]>=0.114.0,<1.0.0",
    "pycryptodome>=3.22.0",
    "python-box>=7.3.2",
    "python-magic>=0.4.27",  # Corrected! Replaced python-magic-bin
    "radon>=5.1.0",
    "z3-solver>=4.14.1.0",
]

[project.scripts]
obfuscation_deobfuscation_crew = "obfuscation_deobfuscation_crew.main:run"
run_crew = "obfuscation_deobfuscation_crew.main:run"
train = "obfuscation_deobfuscation_crew.main:train"
replay = "obfuscation_deobfuscation_crew.main:replay"
test = "obfuscation_deobfuscation_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
